Filters are in separete folder. Style with small changes
