+++ client/src/components/Filters/Filters.css
@@ -1,0 +1,73 @@
+.shows__filter-textcontainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .shows__filter-text {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border: 1px solid white;
+    /* width: 34.9vw; */
+    width: 15.5rem;
+    height: 3.5vh;
+    background-color: var(--color-blue);
+    margin-top: 2vh;
+    color: white;
+    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  }
+  
+  .shows__filter-text button {
+    background-color: inherit;
+    color: var(--color-rose);
+    border: none;
+    text-shadow: inherit;
+    font-size: inherit;
+    font-weight: inherit;
+    display: inherit;
+    cursor: pointer;
+  }
+  
+  .shows__filter-datescontainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 2vh;
+  }
+  
+  .shows__filter-box {
+    /* width: 35vw; */
+    width: 15.5rem;
+    background-color: rgba(0, 0, 0, 0.8);
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+    height: 3.5vh;
+    color: white;
+    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border: 1px solid white;
+}
+  
+  .shows__filter-box button {
+    background-color: transparent;
+    color: white;
+    border: none;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 1rem;
+
+  }
+  
+  .shows__filter-box button:hover {
+    color: var(--color-rose);
+    cursor: pointer;
+  }
+  
+  .shows__filter-box button:active {
+    color: orangered;
+  }
