+++ client/src/components/Shows/Shows.css
@@ -17,77 +17,7 @@
   margin-top: 2vh;
   border: 2px solid var(--color-gray);
 }
-.shows__filter-textcontainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
 
-.shows__filter-text {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid white;
-  /* width: 34.9vw; */
-  width: 14.9rem;
-  height: 3.5vh;
-  background-color: var(--color-blue);
-  margin-top: 2vh;
-  color: white;
-  text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-}
-
-.shows__filter-text button {
-  background-color: inherit;
-  color: var(--color-rose);
-  border: none;
-  text-shadow: inherit;
-  font-size: inherit;
-  font-weight: inherit;
-  display: inherit;
-  cursor: pointer;
-}
-
-.shows__filter-datescontainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 2vh;
-}
-
-.shows__filter-box {
-  /* width: 35vw; */
-  width: 15rem;
-  background-color: rgba(0, 0, 0, 0.8);
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
-  height: 3.5vh;
-  color: white;
-  text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-}
-
-.shows__filter-box button {
-  background-color: inherit;
-  color: inherit;
-  border: 1px solid white;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: medium;
-  cursor: pointer;
-}
-
-.shows__filter-box button:hover {
-  color: var(--color-rose);
-}
-
-.shows__filter-box button:active {
-  color: orangered;
-}
 
 .shows__cards-container {
   display: grid;
@@ -290,31 +220,6 @@ hasta aca */
 
 .shows__categories-buttons:hover {
   color: var(--color-rose);
-}
-
-/*Select*/
-.filter_date_select{
-  width: 100%;
-  height: 1.4rem;
-  background-color: var(--color-blue);
-  color: white;
-  padding-left: 0.2rem;
-  font-size: 0.8rem;
-  font-weight: 600;
-  outline:none; 
-  border-radius: 0 !important;
-}
-
-.filter_date_select option {
-  height: 1.4rem;
-  font-size: 0.8rem;
-  font-weight: 600;
-  margin-bottom: 0.2rem;
-  outline:none; 
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  border-radius: 0 !important;
 }
 
 @media screen and (max-width: 992px) {
